requirements.txt
aiogram>=3.0.0
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
greenlet>=3.0.0
